Catch all exception when looking for main classes
Right now both `ConfigurationContributor` and `TypeSystem` assume that
the classpath given as input is ordered, and that the first entry is a
classes folder and will contain the application classes.

If a main class is not found in the first classpath entry, the next
entry considered is a JAR which fails the entire algorithm. As a first
step, this commit catches such exceptions.

We should consider extracting the main class and main package
information in a different way.